+++ src/main/java/tn/esprit/pidev/consommitounsi/controllers/UserController.java
@@ -1,19 +1,36 @@
 package tn.esprit.pidev.consommitounsi.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import tn.esprit.pidev.consommitounsi.entities.user.User;
 import tn.esprit.pidev.consommitounsi.entities.user.UserErrors;
 import tn.esprit.pidev.consommitounsi.entities.user.UserType;
 import tn.esprit.pidev.consommitounsi.entities.common.Address;
 import tn.esprit.pidev.consommitounsi.services.user.IUserService;
+import tn.esprit.pidev.consommitounsi.utils.UserSecurity;
+import tn.esprit.pidev.consommitounsi.utils.UserSession;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 public class UserController {
     @Autowired
     IUserService userService;
+
+    @PostMapping("/users/login")
+    @ResponseBody
+    public User login() {
+        Authentication auth= SecurityContextHolder.getContext().getAuthentication();
+        if (auth instanceof AnonymousAuthenticationToken || auth == null)
+            return null;
+        User user=userService.getByUsernameOrEmail(auth.getName());
+        UserSession.setUser(user);
+        return user;
+    }
 
     @PostMapping("/users")
     @ResponseBody
@@ -22,84 +39,97 @@ public class UserController {
             return UserErrors.USERNAME_ALREADY_EXISTS;
         if (userService.getByUsernameOrEmail(user.getEmail())!=null)
             return UserErrors.EMAIL_ALREADY_EXISTS;
-        //hash password
+        user.setPassword(UserSecurity.encodePassword(user.getPassword()));
         user.setType(UserType.CUSTOMER);
         userService.addOrUpdate(user);
         return UserErrors.SUCCESS;
     }
 
-    @PostMapping("/users/edit")
+    @PostMapping("/customer/users/edit")
     @ResponseBody
     public UserErrors updateUser(@RequestBody User user) {
-        User existingUser=userService.getByUsernameOrEmail(user.getUsername());
-        if (existingUser!=null && existingUser.getId()!=user.getId())
-            return UserErrors.USERNAME_ALREADY_EXISTS;
-        existingUser=userService.getByUsernameOrEmail(user.getEmail());
-        if (existingUser!=null && existingUser.getId()!=user.getId())
-            return UserErrors.EMAIL_ALREADY_EXISTS;
-        //hash password
-        user.setType(userService.getById(user.getId()).getType());
-        userService.addOrUpdate(user);
-        return UserErrors.SUCCESS;
+        if (UserSession.hasId(user.getId())||UserSession.isAdmin()) {
+            User existingUser = userService.getByUsernameOrEmail(user.getUsername());
+            if (existingUser != null && existingUser.getId() != user.getId())
+                return UserErrors.USERNAME_ALREADY_EXISTS;
+            existingUser = userService.getByUsernameOrEmail(user.getEmail());
+            if (existingUser != null && existingUser.getId() != user.getId())
+                return UserErrors.EMAIL_ALREADY_EXISTS;
+            user.setPassword(UserSecurity.encodePassword(user.getPassword()));
+            user.setType(userService.getById(user.getId()).getType());
+            userService.addOrUpdate(user);
+            return UserErrors.SUCCESS;
+        }
+        return UserErrors.ERROR;
     }
 
-    @PutMapping("/users/{id}/{type}")
+    @PutMapping("/admin/users/{id}/{type}")
     @ResponseBody
     public void updateUserType(@PathVariable("id")long id, @PathVariable("type")UserType type) {
         userService.updateType(id, type);
     }
 
-    @GetMapping("/users/{id}")
+    @GetMapping("/customer/users/{id}")
     @ResponseBody
     public User getUserById(@PathVariable("id")long id) {
         return userService.getById(id);
     }
 
-    @GetMapping("/users/find/{username}")
+    @GetMapping("/customer/users/find/{username}")
     @ResponseBody
     public User getUserByUsernameOrEmail(@PathVariable("username")String username) {
         return userService.getByUsernameOrEmail(username);
     }
 
-    @GetMapping("/users")
+    @GetMapping("/customer/users")
     @ResponseBody
     public List<User> getAllUsers() {
         return userService.getAll();
     }
 
-    @DeleteMapping("/users/{id}")
+    @DeleteMapping("/admin/users/{id}")
     @ResponseBody
     public void deleteUser(@PathVariable("id")long id) {
         userService.delete(id);
     }
 
-    @PostMapping("/addresses/{userId}")
+    @PostMapping("/customer/addresses/{userId}")
     @ResponseBody
     public void addAddress(@RequestBody Address a, @PathVariable("userId") long userId) {
-        userService.addAddress(a, userId);
+        if (UserSession.hasId(userId) || UserSession.isAdmin())
+            userService.addAddress(a, userId);
     }
 
-    @PostMapping("/addresses/edit")
+    @PostMapping("/customer/addresses/edit")
     @ResponseBody
     public void updateAddress(@RequestBody Address a) {
-        userService.updateAddress(a);
+        Address address=userService.getAddressById(a.getId());
+        if (address!=null&&(UserSession.hasId(address.getUser().getId()) || UserSession.isAdmin()))
+            userService.updateAddress(a);
     }
 
-    @GetMapping("/addresses/{id}")
+    @GetMapping("/customer/addresses/{id}")
     @ResponseBody
     public Address getAddressById(@PathVariable("id")long id) {
-        return userService.getAddressById(id);
+        Address a=userService.getAddressById(id);
+        if (a!=null && (UserSession.hasId(a.getUser().getId())||UserSession.isAdmin()||UserSession.isDeliverer()))
+            return a;
+        return null;
     }
 
-    @GetMapping("/users/{userId}/addresses")
+    @GetMapping("/customer/users/{userId}/addresses")
     @ResponseBody
     public List<Address> getUserAddresses(@PathVariable("userId")long userId) {
-        return userService.getUserAddresses(userId);
+        if (UserSession.hasId(userId) || UserSession.isAdmin() || UserSession.isDeliverer())
+            return userService.getUserAddresses(userId);
+        return new ArrayList<>();
     }
 
-    @DeleteMapping("/addresses/{id}")
+    @DeleteMapping("/customer/addresses/{id}")
     @ResponseBody
     public void deleteAddress(@PathVariable("id")long id) {
-        userService.deleteAddressById(id);
+        Address a=userService.getAddressById(id);
+        if (a!=null&&(UserSession.hasId(a.getUser().getId()) || UserSession.isAdmin()))
+            userService.deleteAddressById(id);
     }
 }
